feat: LoanAdvice 기능 구현 및 상품 추가

+++ src/main/java/com/bankersCalculator/server/advice/loanAdvice/dto/request/LoanAdviceRequest.java
@@ -1,8 +1,8 @@
 package com.bankersCalculator.server.advice.loanAdvice.dto.request;
 
-import com.bankersCalculator.server.common.enums.loanAdvise.ChildStatus;
-import com.bankersCalculator.server.common.enums.loanAdvise.MaritalStatus;
-import com.bankersCalculator.server.common.enums.ltv.HouseOwnershipType;
+import com.bankersCalculator.server.common.enums.loanAdvice.ChildStatus;
+import com.bankersCalculator.server.common.enums.loanAdvice.MaritalStatus;
+import com.bankersCalculator.server.common.enums.calculator.HouseOwnershipType;
 import com.bankersCalculator.server.housingInfo.rentTransactionInquiry.common.RentHousingType;
 import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.NotNull;
@@ -58,7 +58,7 @@ public class LoanAdviceRequest {
     // 주택정보
     private RentHousingType rentHousingType;  // 주택타입
 
-    private Double exclusiveArea; // 전용면적
+    private BigDecimal exclusiveArea; // 전용면적
 
     private String buildingName;   // 건물명
