+++ src/main/java/com/bankersCalculator/server/advice/userInputInfo/service/UserInputInfoService.java
@@ -2,12 +2,13 @@ package com.bankersCalculator.server.advice.userInputInfo.service;
 
 import com.bankersCalculator.server.advice.userInputInfo.dto.UserInputInfoResponse;
 import com.bankersCalculator.server.advice.userInputInfo.dto.UserInputSummaryResponse;
-import com.bankersCalculator.server.common.enums.loanAdvise.ChildStatus;
-import com.bankersCalculator.server.common.enums.loanAdvise.MaritalStatus;
+import com.bankersCalculator.server.common.enums.loanAdvice.ChildStatus;
+import com.bankersCalculator.server.common.enums.loanAdvice.MaritalStatus;
 import com.bankersCalculator.server.housingInfo.rentTransactionInquiry.common.RentHousingType;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @RequiredArgsConstructor
@@ -17,19 +18,19 @@ public class UserInputInfoService {
     public UserInputInfoResponse getRecentlySubmittedUserInput() {
 
         return UserInputInfoResponse.builder()
-            .rentalDeposit(300000000L)  // 3억원 임차보증금
-            .monthlyRent(500000L)       // 50만원 월세
-            .cashOnHand(50000000L)      // 5천만원 보유 현금
+            .rentalDeposit(BigDecimal.valueOf(300000000))  // 3억원 임차보증금
+            .monthlyRent(BigDecimal.valueOf(500000))       // 50만원 월세
+            .cashOnHand(BigDecimal.valueOf(50000000))      // 5천만원 보유 현금
             .age(35)                    // 35세
             .maritalStatus(MaritalStatus.MARRIED)
-            .annualIncome(60000000L)    // 6천만원 연소득
-            .spouseAnnualIncome(40000000L)  // 4천만원 배우자 연소득
+            .annualIncome(BigDecimal.valueOf(60000000))    // 6천만원 연소득
+            .spouseAnnualIncome(BigDecimal.valueOf(40000000))  // 4천만원 배우자 연소득
             .childStatus(ChildStatus.ONE_CHILD)
             .hasNewborn(true)
             .isSMEEmployee(true)        // 중소기업 재직 여부
             .isNetAssetOver345M(false)  // 순자산 3.45억 초과 여부
             .rentHousingType(RentHousingType.APARTMENT)
-            .exclusiveArea(85.0)         // 85제곱미터 전용면적
+            .exclusiveArea(BigDecimal.valueOf(85.0))         // 85제곱미터 전용면적
             .buildingName("행복아파트")
             .districtCode("1168010100") // 서울특별시 강남구 삼성동
             .dongName("삼성동")
@@ -40,19 +41,19 @@ public class UserInputInfoService {
     public UserInputInfoResponse getSpecificUserInput(Long userInfoInputId) {
 
         return UserInputInfoResponse.builder()
-            .rentalDeposit(300000000L)  // 3억원 임차보증금
-            .monthlyRent(500000L)       // 50만원 월세
-            .cashOnHand(50000000L)      // 5천만원 보유 현금
+            .rentalDeposit(BigDecimal.valueOf(300000000))  // 3억원 임차보증금
+            .monthlyRent(BigDecimal.valueOf(500000))       // 50만원 월세
+            .cashOnHand(BigDecimal.valueOf(50000000))      // 5천만원 보유 현금
             .age(35)                    // 35세
             .maritalStatus(MaritalStatus.MARRIED)
-            .annualIncome(60000000L)    // 6천만원 연소득
-            .spouseAnnualIncome(40000000L)  // 4천만원 배우자 연소득
+            .annualIncome(BigDecimal.valueOf(60000000))    // 6천만원 연소득
+            .spouseAnnualIncome(BigDecimal.valueOf(40000000))  // 4천만원 배우자 연소득
             .childStatus(ChildStatus.ONE_CHILD)
             .hasNewborn(true)
             .isSMEEmployee(true)        // 중소기업 재직 여부
             .isNetAssetOver345M(false)  // 순자산 3.45억 초과 여부
             .rentHousingType(RentHousingType.APARTMENT)
-            .exclusiveArea(85.0)         // 85제곱미터 전용면적
+            .exclusiveArea(BigDecimal.valueOf(85.0))         // 85제곱미터 전용면적
             .buildingName("행복아파트")
             .districtCode("1168010100") // 서울특별시 강남구 삼성동
             .dongName("삼성동")
